+++ src/main/java/com/example/kkubeurakko/global/config/WebClientConfig.java
@@ -1,4 +1,4 @@
-package com.example.kkubeurakko.global;
+package com.example.kkubeurakko.global.config;
 
 import io.netty.channel.ChannelOption;
 import io.netty.handler.timeout.ReadTimeoutHandler;
