+++ src/main/java/com/example/kkubeurakko/global/controller/KakaoLoginController.java
@@ -1,0 +1,2 @@
+package com.example.kkubeurakko.global.controller;public class KakaoLoginController {
+}
