+++ src/main/java/com/example/kkubeurakko/global/dto/KakaoTokenResponseDto.java
@@ -1,0 +1,2 @@
+package com.example.kkubeurakko.global.dto;public class KakaoTokenResponseDto {
+}
