+++ src/main/java/com/example/kkubeurakko/global/service/KakaoService.java
@@ -1,0 +1,2 @@
+package com.example.kkubeurakko.global.service;public class KakaoService {
+}
